pyproject.toml
[project]
name = "dental-appointement-booking-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.26",
    "langchain-openai>=1.1.12",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.6",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "streamlit>=1.56.0",
]
